feat(taskEitherChainTap) - introduce function to chain and return previous values
