+++ packages/fp-ts-extensions/src/lib/fp-ts-extensions.spec.ts
@@ -3,29 +3,7 @@ import '@relmify/jest-fp-ts';
 import { taskEitherChainTap } from './fp-ts-extensions';
 import { either, taskEither } from 'fp-ts';
 import { pipe } from 'fp-ts/function';
-
-export class GetProjectionError extends Error {
-  public readonly _tag: 'GetProjectionError';
-
-  private constructor() {
-    super('failed to get projection');
-    this._tag = 'GetProjectionError';
-  }
-
-  public static readonly of = (): GetProjectionError =>
-    new GetProjectionError();
-}
-
-export class AnotherError extends Error {
-  public readonly _tag: 'AnotherError';
-
-  private constructor() {
-    super('something bad happened');
-    this._tag = 'AnotherError';
-  }
-
-  public static readonly of = (): AnotherError => new AnotherError();
-}
+import {AnotherError, GetProjectionError} from "../fixtures/errors";
 
 describe('fp-ts extensions', () => {
   describe('taskEitherChainTap', () => {
